[minor] Add openapi docs, refactor some routing
