+++ preboots/swagger.js
@@ -1,0 +1,6 @@
+const swaggerDocumentation = require('../swagger/documentation.json');
+
+module.exports = function (app, options, done) {
+  app.swaggerDocumentation = swaggerDocumentation;
+  done();
+};
